# api.yaml
...
openapi: 3.0.0
paths:
...
  "/nft":
    post:
      operationId: NftController_mint
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateNftDto"
      responses:
        '201':
          description: ''
      tags:
      - NFT
      security:
      - bearer: []
    get:
      operationId: NftController_getTransactionByAction
      parameters: []
      responses:
        '200':
          description: ''
      tags:
      - NFT
      security:
      - bearer: []
  "/collection":
    post:
      operationId: CollectionController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateCollectionDto"
      responses:
        '201':
          description: ''
      tags:
      - Collection
      security:
      - bearer: []
    get:
      operationId: CollectionController_getTransactionByAction
      parameters: []
      responses:
        '200':
          description: ''
      tags:
      - Collection
      security:
      - bearer: []
...
  "/list":
    post:
      operationId: ListController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateListDto"
      responses:
        '201':
          description: ''
      tags:
      - List
      security:
      - bearer: []
    get:
      operationId: ListController_getTransactionByAction
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/GetListDto"
      responses:
        '200':
          description: ''
      tags:
      - List
      security:
      - bearer: []
  "/list/account":
    get:
      operationId: ListController_getByAccount
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ListByAccountDto"
      responses:
        '200':
          description: ''
      tags:
      - List
      security:
      - bearer: []
  "/list/{id}":
    patch:
      operationId: ListController_updateList
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateListDto"
      responses:
        '200':
          description: ''
      tags:
      - List
      security:
      - bearer: []
    delete:
      operationId: ListController_deleteList
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: ''
      tags:
      - List
      security:
      - bearer: []
  "/loan":
    post:
      operationId: LoanController_mint
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LoanDto"
      responses:
        '201':
          description: ''
      tags:
      - Loan
      security:
      - bearer: []
    get:
      operationId: LoanController_getTransactionByAction
      parameters: []
      responses:
        '200':
          description: ''
      tags:
      - Loan
      security:
      - bearer: []
  "/loan/{id}":
    patch:
      operationId: LoanController_updateLoan
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateLoanDto"
      responses:
        '200':
          description: ''
      tags:
      - Loan
      security:
      - bearer: []
  "/loan/:id":
    delete:
      operationId: LoanController_deleteLoan
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: ''
      tags:
      - Loan
      security:
      - bearer: []
...
  "/history":
    post:
      operationId: HistoryController_mint
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateHistoryDto"
      responses:
        '201':
          description: ''
      tags:
      - History
      security:
      - bearer: []
    get:
      operationId: HistoryController_getTransactionByAction
      parameters: []
      responses:
        '200':
          description: ''
      tags:
      - History
      security:
      - bearer: []
...
  "/notification":
    post:
      operationId: NotificationController_mint
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateNotificationDto"
      responses:
        '201':
          description: ''
      tags:
      - Notification
      security:
      - bearer: []
    get:
      operationId: NotificationController_getTransactionByAction
      parameters: []
      responses:
        '200':
          description: ''
      tags:
      - Notification
      security:
      - bearer: []
  "/loanHistory":
    post:
      operationId: LoanHistoryController_mint
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateLoanHistoryDto"
      responses:
        '201':
          description: ''
      tags:
      - LoanHistory
      security:
      - bearer: []
    get:
      operationId: LoanHistoryController_getTransactionByAction
      parameters: []
      responses:
        '200':
          description: ''
      tags:
      - LoanHistory
      security:
      - bearer: []
  "/listHistory":
    post:
      operationId: ListHistoryController_mint
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateListHistoryDto"
      responses:
        '201':
          description: ''
      tags:
      - ListHistory
      security:
      - bearer: []
    get:
      operationId: ListHistoryController_getTransactionByAction
      parameters: []
      responses:
        '200':
          description: ''
      tags:
      - ListHistory
      security:
      - bearer: []
...
info:
  title: Solarr API Documentation
  description: Solarr API Documentation
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateNftDto:
      type: object
      properties:
        name:
          type: string
        tokenId:
          type: number
        state:
          type: number
        metadata:
          type: string
        collectionFloor:
          type: number
        bankValuation:
          type: number
        ownerId:
          type: number
        collectionId:
          type: number
      required:
      - name
      - tokenId
      - state
      - metadata
      - collectionFloor
      - bankValuation
      - ownerId
      - collectionId
    CreateCollectionDto:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
      required:
      - name
      - address
    CreateListDto:
      type: object
      properties:
        loanAmount:
          type: number
        loanType:
          type: boolean
        period:
          type: number
        apr:
          type: number
        borrowerId:
          type: number
        collectionAddress:
          type: string
        tokenId:
          type: number
        metadata:
          type: string
      required:
      - loanAmount
      - loanType
      - period
      - apr
      - borrowerId
      - collectionAddress
      - tokenId
      - metadata
    GetListDto:
      type: object
      properties:
        lenderId:
          type: number
      required:
      - lenderId
    ListByAccountDto:
      type: object
      properties:
        borrowerId:
          type: number
        nfts:
          type: array
          items:
            type: string
      required:
      - borrowerId
      - nfts
    UpdateListDto:
      type: object
      properties:
        loanAmount:
          type: number
        loanType:
          type: boolean
        period:
          type: number
        apr:
          type: number
      required:
      - loanAmount
      - loanType
      - period
      - apr
    LoanDto:
      type: object
      properties:
        state:
          type: number
        borrowerId:
          type: number
        lenderId:
          type: number
        collectionAddress:
          type: string
        tokenId:
          type: number
        acceptedOfferId:
          type: number
      required:
      - state
      - borrowerId
      - lenderId
      - collectionAddress
      - tokenId
      - acceptedOfferId
    UpdateLoanDto:
      type: object
      properties:
        state:
          type: number
      required:
      - state
...
    CreateHistoryDto:
      type: object
      properties:
        type:
          type: number
        borrowerAction:
          type: number
        lenderAction:
          type: number
        userId:
          type: number
        nftId:
          type: number
      required:
      - type
      - borrowerAction
      - lenderAction
      - userId
      - nftId
...
    CreateNotificationDto:
      type: object
      properties:
        userId:
          type: number
        nftId:
          type: number
        status:
          type: number
      required:
      - userId
      - nftId
      - status
    CreateLoanHistoryDto:
      type: object
      properties:
        action:
          type: number
        borrowerId:
          type: number
        lenderId:
          type: number
        loanId:
          type: number
      required:
      - action
      - borrowerId
      - lenderId
      - loanId
    CreateListHistoryDto:
      type: object
      properties:
        action:
          type: number
        listId:
          type: number
      required:
      - action
      - listId
...